Remove old editor hook
