+++ packages/block-editor/src/components/block-list/use-block-props/index.js
@@ -30,7 +30,7 @@ import { useBlockRefProvider } from './use-block-refs';
 import { useIntersectionObserver } from './use-intersection-observer';
 import { useScrollIntoView } from './use-scroll-into-view';
 import { useFlashEditableBlocks } from '../../use-flash-editable-blocks';
-import { canBindBlock } from '../../../hooks/use-bindings-attributes';
+import { canBindBlock } from '../../../utils/bindings';
 
 /**
  * This hook is used to lightly mark an element as a block element. The element
