+++ packages/block-editor/src/components/rich-text/index.js
@@ -38,7 +38,7 @@ import { getAllowedFormats } from './utils';
 import { Content, valueToHTMLString } from './content';
 import { withDeprecations } from './with-deprecations';
 import { unlock } from '../../lock-unlock';
-import { canBindBlock } from '../../hooks/use-bindings-attributes';
+import { canBindBlock } from '../../utils/bindings';
 
 export const keyboardShortcutContext = createContext();
 export const inputEventContext = createContext();
