+++ packages/block-editor/src/hooks/index.js
@@ -31,7 +31,6 @@ import './metadata';
 import blockHooks from './block-hooks';
 import blockBindingsPanel from './block-bindings';
 import './block-renaming';
-import './use-bindings-attributes';
 
 createBlockEditFilter(
 	[
